requirements.txt
requests
pyyaml
pandas
websockets
schedule
yfinance
praw
beautifulsoup4
lxml
anthropic
google-generativeai
